refactor(social): eliminate useEffect in channels and layout
Apply Vercel React best practices across channels + layout:
- channels/page.tsx: Server Component reads searchParams, normalizes
  OAuth callback params, passes to ChannelsPageClient
- ChannelsPageClient: processes OAuth callback via ref init (no useEffect)
- PlatformPicker: module-level provider cache (no useEffect fetch)
- SocialLayout: ref-based account fetch init (no useEffect)

Only 2 useEffect calls remain in entire Social Hub — both are
legitimate DOM operations (click-outside listener, textarea auto-grow).
